# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "tinrist123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/category/post/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get Post On Category By Id",
                "description": "Return Posts",
                "operationId": "getPostOnCateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/category/post/total/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get Total Post By Category",
                "description": "Returns project number of Post",
                "operationId": "getTotalPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:true",
                            "read:true"
                        ]
                    }
                ]
            }
        },
        "/api/v1/category/post/{id}/page={page}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get Post On Category By Id Per Page",
                "description": "Return Posts Per Page",
                "operationId": "getProjectByIdAndPage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/category/all": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of Category",
                "operationId": "getAllCate",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/v1/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get list of Category",
                "description": "Returns list of Category",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}